+++ erpnext/apps/rootedops_payroll/rootedops_payroll/install.py
@@ -1,4 +1,5 @@
 from rootedops_payroll.services.employee_payments import ensure_employee_payment_custom_fields
+from rootedops_payroll.services.plaid_item import ensure_plaid_item_custom_fields
 from rootedops_payroll.services.payroll_engine import ensure_employee_tax_profile_custom_fields
 from rootedops_payroll.services.tax_compliance import ensure_tax_payment_custom_fields
 
@@ -7,3 +8,4 @@ def after_migrate():
     ensure_employee_tax_profile_custom_fields()
     ensure_employee_payment_custom_fields()
     ensure_tax_payment_custom_fields()
+    ensure_plaid_item_custom_fields()